- Fixed an issue with the possible angle of the clock
- Added the multiplier to the tables
 - Has to change the score from int to float to be able to multiple it

+++ Assets/LAGS/Scripts/Pub/PubManager.cs
@@ -194,7 +194,7 @@ namespace LAGS.Managers.Pub
 
         #region Score
 
-        public void AddScore(int score)
+        public void AddScore(float score)
         {
             _currentScore += score;
         }
